+++ src/com/google/android/diskusage/FileSystemSystemSpace.java
@@ -1,7 +1,7 @@
 package com.google.android.diskusage;
 
 public class FileSystemSystemSpace extends FileSystemSpecial {
-  public FileSystemSystemSpace(String name, long size, long blockSize) {
+  public FileSystemSystemSpace(String name, long size, int blockSize) {
     super(name, size, blockSize);
   }
 }
